refactoring part 1
