+++ ogcapi-stable/ogcapi-html/src/main/javascript/src/components/MapLibre/LayerControl/EntriesLayer.jsx
@@ -1,0 +1,131 @@
+import React, { useState } from "react";
+import PropTypes from "prop-types";
+import { Collapse } from "reactstrap";
+import SubLayers from "./SubLayers";
+
+const Entries = ({
+  p,
+  isSubLayerOpen,
+  selected,
+  selectedBasemap,
+  setSelected,
+  setSelectedBasemap,
+  allParentGroups,
+  open,
+  setOpen,
+}) => {
+  const onSelectEntry = (entry) => {
+    const subLayers = entry.subLayers.map((subLayer) => {
+      return subLayer.id;
+    });
+    if (subLayers.every((id) => selected.includes(id))) {
+      setSelected(selected.filter((id) => !subLayers.includes(id)));
+    } else {
+      setSelected([...selected, ...subLayers]);
+    }
+  };
+
+  const onSelectRadio = (entry) => {
+    if (!selectedBasemap.includes(entry.id)) {
+      setSelectedBasemap(null);
+      setSelectedBasemap([entry.id]);
+    }
+  };
+
+  const onOpen = (entry) => {
+    const index = open.indexOf(entry.id);
+    if (index < 0) {
+      open.push(entry.id);
+    } else {
+      open.splice(index, 1);
+    }
+    setOpen([...open]);
+  };
+
+  return (
+    <>
+      {allParentGroups
+        ? p.entries.map((entry) => (
+            <div key={entry.id}>
+              <Collapse
+                isOpen={isSubLayerOpen(p.id)}
+                id={`collapse-${entry.id}`}
+                key={entry.id}
+                className="accordion-collapse"
+                aria-labelledby={`heading-${entry.id}`}
+                data-bs-parent="#layer-control"
+              >
+                {p.isBasemap !== true ? (
+                  <div>
+                    <button
+                      style={{
+                        backgroundColor: "white",
+                        borderRadius: "0.25rem",
+                        padding: "10px",
+                      }}
+                      color="secondary"
+                      outline
+                      onClick={(e) => {
+                        if (!e.target.classList.contains("form-check-input")) {
+                          e.target.blur();
+                          onOpen(entry);
+                        }
+                      }}
+                      active={isSubLayerOpen(entry.id)}
+                      className={`accordion-button ${isSubLayerOpen(entry.id) ? "collapsed" : ""}`}
+                      type="button"
+                      data-bs-toggle="collapse"
+                      data-bs-target={`#collapse-${entry.id}`}
+                      aria-expanded={isSubLayerOpen(entry.id)}
+                      aria-controls={`collapse-${entry.id}`}
+                    >
+                      <input
+                        style={{ marginRight: "5px" }}
+                        className="form-check-input"
+                        type="checkbox"
+                        id={`checkbox-${entry.id}`}
+                        checked={entry.subLayers.every((subLayer) =>
+                          selected.includes(subLayer.id)
+                        )}
+                        onChange={(e) => {
+                          e.target.blur();
+                          onSelectEntry(entry);
+                        }}
+                      />
+
+                      <span style={{ marginRight: "10px" }}>{entry.id}</span>
+                    </button>
+                  </div>
+                ) : (
+                  <div>
+                    <input
+                      style={{ margin: "10px" }}
+                      className="form-check-input"
+                      type="radio"
+                      id={`radiobutton-${entry.id}`}
+                      name="basemap"
+                      value={`${entry.id}`}
+                      checked={selectedBasemap.includes(entry.id)}
+                      onClick={(e) => {
+                        e.target.blur();
+                        onSelectRadio(entry);
+                      }}
+                    />
+                    <span style={{ marginRight: "10px" }}>{entry.id}</span>
+                  </div>
+                )}
+              </Collapse>
+              <SubLayers
+                entry={entry}
+                isSubLayerOpen={isSubLayerOpen}
+                selected={selected}
+                setSelected={setSelected}
+              />
+            </div>
+          ))
+        : null}
+    </>
+  );
+};
+
+export default Entries;
