+++ ogcapi-stable/ogcapi-html/src/main/javascript/src/components/MapLibre/LayerControl/index.jsx
@@ -1,8 +1,8 @@
 import React, { useState } from "react";
 import PropTypes from "prop-types";
-import { Collapse } from "reactstrap";
 import { useMaplibreUIEffect } from "react-maplibre-ui";
 import { useMapVisibility } from "./fetchStyle";
+import Entries from "./EntriesLayer";
 
 const LayerControl = ({ layerGroups }) => {
   const parent = layerGroups.filter((entry) => entry.type === "group");
@@ -48,34 +48,6 @@ const LayerControl = ({ layerGroups }) => {
     return subLayerIds.every((id) => selected.includes(id));
   };
 
-  const onSelectEntry = (entry) => {
-    const subLayers = entry.subLayers.map((subLayer) => {
-      return subLayer.id;
-    });
-    if (subLayers.every((id) => selected.includes(id))) {
-      setSelected(selected.filter((id) => !subLayers.includes(id)));
-    } else {
-      setSelected([...selected, ...subLayers]);
-    }
-  };
-
-  const onSelect = (entry) => {
-    const index = selected.indexOf(entry.id);
-    if (index < 0) {
-      selected.push(entry.id);
-    } else {
-      selected.splice(index, 1);
-    }
-    setSelected([...selected]);
-  };
-
-  const onSelectRadio = (entry) => {
-    if (!selectedBasemap.includes(entry.id)) {
-      setSelectedBasemap(null);
-      setSelectedBasemap([entry.id]);
-    }
-  };
-
   const onOpenParent = (entry) => {
     if (entry.isBasemap !== true) {
       const entryIds = [entry.id, ...entry.entries.map((e) => e.id)];
@@ -98,16 +70,6 @@ const LayerControl = ({ layerGroups }) => {
         setOpen(open.filter((ids) => !entry.id.includes(ids)));
       }
     }
-  };
-
-  const onOpen = (entry) => {
-    const index = open.indexOf(entry.id);
-    if (index < 0) {
-      open.push(entry.id);
-    } else {
-      open.splice(index, 1);
-    }
-    setOpen([...open]);
   };
 
   useMaplibreUIEffect(({ map }) => {
@@ -139,16 +101,6 @@ const LayerControl = ({ layerGroups }) => {
   const isSubLayerOpen = (name) => {
     return open.includes(name);
   };
-  /*
-  //   Test von allParentGroups und parent.id:
-  console.log("allPGroups", allParentGroups);
-  console.log(
-    "parent",
-    parent.map((p) => {
-      return p;
-    })
-  );
-*/
 
   return (
     <>
@@ -246,107 +198,17 @@ const LayerControl = ({ layerGroups }) => {
                 </button>
               </h2>
 
-              {allParentGroups
-                ? p.entries.map((entry) => (
-                    <div key={entry.id}>
-                      <Collapse
-                        isOpen={isSubLayerOpen(p.id)}
-                        id={`collapse-${entry.id}`}
-                        key={entry.id}
-                        className="accordion-collapse"
-                        aria-labelledby={`heading-${entry.id}`}
-                        data-bs-parent="#layer-control"
-                      >
-                        {p.isBasemap !== true ? (
-                          <div>
-                            <button
-                              style={{
-                                backgroundColor: "white",
-                                borderRadius: "0.25rem",
-                                padding: "10px",
-                              }}
-                              color="secondary"
-                              outline
-                              onClick={(e) => {
-                                if (!e.target.classList.contains("form-check-input")) {
-                                  e.target.blur();
-                                  onOpen(entry);
-                                }
-                              }}
-                              active={isSubLayerOpen(entry.id)}
-                              className={`accordion-button ${
-                                isSubLayerOpen(entry.id) ? "collapsed" : ""
-                              }`}
-                              type="button"
-                              data-bs-toggle="collapse"
-                              data-bs-target={`#collapse-${entry.id}`}
-                              aria-expanded={isSubLayerOpen(entry.id)}
-                              aria-controls={`collapse-${entry.id}`}
-                            >
-                              <input
-                                style={{ marginRight: "5px" }}
-                                className="form-check-input"
-                                type="checkbox"
-                                id={`checkbox-${entry.id}`}
-                                checked={entry.subLayers.every((subLayer) =>
-                                  selected.includes(subLayer.id)
-                                )}
-                                onChange={(e) => {
-                                  e.target.blur();
-                                  onSelectEntry(entry);
-                                }}
-                              />
-
-                              <span style={{ marginRight: "10px" }}>{entry.id}</span>
-                            </button>
-                          </div>
-                        ) : (
-                          <div>
-                            <input
-                              style={{ margin: "10px" }}
-                              className="form-check-input"
-                              type="radio"
-                              id={`radiobutton-${entry.id}`}
-                              name="basemap"
-                              value={`${entry.id}`}
-                              checked={selectedBasemap.includes(entry.id)}
-                              onClick={(e) => {
-                                e.target.blur();
-                                onSelectRadio(entry);
-                              }}
-                            />
-                            <span style={{ marginRight: "10px" }}>{entry.id}</span>
-                          </div>
-                        )}
-                      </Collapse>
-
-                      {entry.subLayers && entry.subLayers.length > 0
-                        ? entry.subLayers.map((subLayer) => (
-                            <div key={subLayer.id}>
-                              <Collapse
-                                isOpen={isSubLayerOpen(entry.id)}
-                                id={`collapse-${subLayer.id}`}
-                                key={subLayer.id}
-                                className="accordion-collapse"
-                                aria-labelledby={`heading-${subLayer.id}`}
-                                data-bs-parent="#layer-control"
-                              >
-                                <input
-                                  style={{ marginLeft: "5px" }}
-                                  className="form-check-input"
-                                  type="checkbox"
-                                  id={`checkbox-${subLayer.id}`}
-                                  checked={selected.includes(subLayer.id)}
-                                  onChange={() => onSelect(subLayer)}
-                                />
-                                <span style={{ marginLeft: "10px" }}>{subLayer.id}</span>
-                              </Collapse>
-                            </div>
-                          ))
-                        : null}
-                    </div>
-                  ))
-                : null}
+              <Entries
+                p={p}
+                isSubLayerOpen={isSubLayerOpen}
+                selected={selected}
+                selectedBasemap={selectedBasemap}
+                setSelected={setSelected}
+                setSelectedBasemap={setSelectedBasemap}
+                allParentGroups={allParentGroups}
+                open={open}
+                setOpen={setOpen}
+              />
             </div>
           ) : null
         )}
